refactor(db) 9/13: extract db/{conquest,schematic,stats_systems}.rs (closes #255)
Move conquest (3), schematic (2), and armor/stat-curve (2) populators + their
classifiers (parse_conquest_fqn, profession_from_fqn), CREATE TABLEs, and tests
into three domain files. mod.rs now 1042 lines (infra only). 240 tests + clippy
-D warnings + fmt green.
