+++ kessel/src/db/mod.rs
@@ -12,10 +12,13 @@ pub(crate) use crate::grammar::Grammar;
 pub(crate) use crate::schema::{decode_payload_schema_aware, GameObject};
 pub(crate) use crate::stb::StbEntry;
 mod ability;
+mod conquest;
 mod item;
 mod npc;
 mod quest;
 mod schema;
+mod schematic;
+mod stats_systems;
 mod world;
 pub(crate) use world::find_planet_code_after_cc;
 #[cfg(test)]
@@ -25,27 +28,6 @@ pub(crate) use util::*;
 
 /// Quest class type_hi32 from client.gom (decoded by Agent D, legion 019e4d75).
 const QUEST_CLASS_TYPE_HI32: u32 = 0x2ADE_C3D2;
-
-/// Recognize a SWTOR crafting profession from any segment of an FQN.
-/// Returns the lowercased profession name when found, None otherwise. Used
-/// by `populate_schematic_details_typed` (#178).
-fn profession_from_fqn(fqn: &str) -> Option<String> {
-    const PROFESSIONS: &[&str] = &[
-        "artifice",
-        "armormech",
-        "armstech",
-        "biochem",
-        "cybertech",
-        "synthweaving",
-    ];
-    let lower = fqn.to_lowercase();
-    for prof in PROFESSIONS {
-        if lower.contains(prof) {
-            return Some((*prof).to_string());
-        }
-    }
-    None
-}
 
 /// Serialized object ready for batch insert
 struct PendingObject {
@@ -174,6 +156,9 @@ impl Database {
             let mut conn = self.conn.lock().unwrap();
             let tx = conn.transaction()?;
             schema::create_tables(&tx)?;
+            stats_systems::create_tables(&tx)?;
+            schematic::create_tables(&tx)?;
+            conquest::create_tables(&tx)?;
             world::create_tables(&tx)?;
             npc::create_tables(&tx)?;
             ability::create_tables(&tx)?;
@@ -560,468 +545,9 @@ impl Database {
 
         Ok(updated)
     }
-
-    /// Populate `schematics` and `schematic_materials` from `itm.schem.*` +
-    /// `schem.*` payloads.
-    ///
-    /// Each `itm.schem.*` object's payload carries a CF GUID ref to a
-    /// companion `schem.*` object (different GOM kind, ~14k instances). The
-    /// schem.* payload encodes the recipe: a list of CF GUID refs each
-    /// followed by a quantity byte. Resolved FQNs are split by prefix:
-    /// `itm.mat.*` rows go to `schematic_materials`, anything else is treated
-    /// as the output and stored in `schematics.output_fqn`.
-    ///
-    /// The quantity byte sits immediately after each 9-byte CF marker
-    /// (`CF E0 NN NN NN NN NN NN NN`). Material values run 1-99 in observed
-    /// payloads (low-bit-set non-CF bytes); the parser clamps to 0..99 to
-    /// reject obviously-non-quantity bytes.
-    pub fn populate_schematic_recipes(&self) -> Result<u64> {
-        use base64::{engine::general_purpose::STANDARD as BASE64, Engine as _};
-        use std::collections::HashMap;
-
-        let conn = self.conn.lock().unwrap();
-
-        // Build GUID -> FQN map for all objects (only need one lookup table).
-        let mut guid_to_fqn: HashMap<String, String> = HashMap::new();
-        {
-            let mut stmt = conn.prepare("SELECT guid, fqn FROM objects")?;
-            for row in stmt
-                .query_map([], |row| {
-                    Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
-                })?
-                .filter_map(|r| r.ok())
-            {
-                guid_to_fqn.insert(row.0.to_uppercase(), row.1);
-            }
-        }
-
-        // Map itm.schem.<X> -> schem.<X> via the strip-prefix convention,
-        // resolved by FQN match (cheap and reliable; the CF ref out of the
-        // itm.schem.* payload would also work but adds a dump pass).
-        // Build schem.* fqn -> payload_b64 map (single scan, indexed lookup).
-        let schem_payloads: HashMap<String, String> = {
-            let mut stmt = conn.prepare(
-                "SELECT fqn, json_extract(json, '$.payload_b64') \
-                 FROM objects WHERE kind = 'schem' AND is_canonical = 1",
-            )?;
-            let collected: HashMap<String, String> = stmt
-                .query_map([], |row| {
-                    Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
-                })?
-                .filter_map(|r| r.ok())
-                .collect();
-            collected
-        };
-
-        // Pair each itm.schem.* with its schem.* companion via the strip-prefix
-        // convention. In-memory map lookup avoids the quadratic SQL JOIN that
-        // would otherwise run REPLACE() against every row pair.
-        let itm_to_schem: Vec<(String, String)> = {
-            let mut stmt = conn.prepare(
-                "SELECT fqn FROM objects WHERE fqn LIKE 'itm.schem.%' AND kind = 'Item' AND is_canonical = 1",
-            )?;
-            let collected: Vec<(String, String)> = stmt
-                .query_map([], |row| row.get::<_, String>(0))?
-                .filter_map(|r| r.ok())
-                .filter_map(|itm_fqn| {
-                    let schem_fqn = itm_fqn.replacen("itm.schem.", "schem.", 1);
-                    schem_payloads.get(&schem_fqn).map(|p| (itm_fqn, p.clone()))
-                })
-                .collect();
-            collected
-        };
-
-        let tx = conn.unchecked_transaction()?;
-        let mut schem_stmt = tx.prepare_cached(
-            "INSERT OR REPLACE INTO schematics (schematic_fqn, output_fqn, output_resolved) \
-             VALUES (?1, ?2, ?3)",
-        )?;
-        let mut mat_stmt = tx.prepare_cached(
-            "INSERT OR IGNORE INTO schematic_materials (schematic_fqn, material_fqn, quantity) \
-             VALUES (?1, ?2, ?3)",
-        )?;
-
-        let mut count = 0u64;
-        for (schematic_fqn, payload_b64) in &itm_to_schem {
-            let Ok(payload) = BASE64.decode(payload_b64) else {
-                continue;
-            };
-
-            let mut output_fqn: Option<String> = None;
-            let mut materials: Vec<(String, u32)> = Vec::new();
-
-            let mut i = 0;
-            while i + 10 <= payload.len() {
-                if payload[i] == 0xCF && payload[i + 1] == 0xE0 {
-                    let ref_guid: String = payload[i + 1..i + 9]
-                        .iter()
-                        .map(|b| format!("{:02X}", b))
-                        .collect();
-                    let qty_byte = payload[i + 9];
-                    if let Some(fqn) = guid_to_fqn.get(&ref_guid) {
-                        if fqn.starts_with("itm.mat.") {
-                            // Quantity follows the 9-byte CF marker. Reject
-                            // values >99 to avoid mistaking a continuation
-                            // byte for a quantity.
-                            let qty = if qty_byte == 0 || qty_byte > 99 {
-                                1
-                            } else {
-                                qty_byte as u32
-                            };
-                            materials.push((fqn.clone(), qty));
-                        } else if fqn.starts_with("itm.")
-                            && !fqn.starts_with("itm.schem.")
-                            && fqn != schematic_fqn
-                            && output_fqn.is_none()
-                        {
-                            output_fqn = Some(fqn.clone());
-                        }
-                    }
-                    i += 9;
-                } else {
-                    i += 1;
-                }
-            }
-
-            let resolved = output_fqn.is_some() as i32;
-            schem_stmt.execute(params![schematic_fqn, output_fqn, resolved])?;
-            count += 1;
-            for (mat_fqn, qty) in &materials {
-                mat_stmt.execute(params![schematic_fqn, mat_fqn, qty])?;
-            }
-        }
-
-        drop(schem_stmt);
-        drop(mat_stmt);
-        tx.commit()?;
-        Ok(count)
-    }
-
-    /// Populate `schematic_details` with FQN-derived profession (#178).
-    ///
-    /// Walks every `schem.*` canonical object, looks for a recognized
-    /// crafting profession token anywhere in the FQN, and records it.
-    /// `tier` and `training_cost` remain NULL pending the per-property
-    /// byte-layout decode work (the int8/16/32/enum_ref/string decode
-    /// gap documented in CLAUDE.md).
-    pub fn populate_schematic_details_typed(&self) -> Result<u64> {
-        self.flush()?;
-        let conn = self.conn.lock().unwrap();
-        let fqns: Vec<String> = {
-            let mut stmt = conn
-                .prepare("SELECT fqn FROM objects WHERE fqn LIKE 'schem.%' AND is_canonical = 1")?;
-            let collected: Vec<String> = stmt
-                .query_map([], |row| row.get::<_, String>(0))?
-                .filter_map(|r| r.ok())
-                .collect();
-            collected
-        };
-        let tx = conn.unchecked_transaction()?;
-        let mut insert = tx.prepare_cached(
-            "INSERT OR REPLACE INTO schematic_details \
-               (fqn, profession, tier, training_cost) \
-             VALUES (?1, ?2, ?3, ?4)",
-        )?;
-        let mut written = 0u64;
-        for fqn in &fqns {
-            let profession = profession_from_fqn(fqn);
-            insert.execute(params![fqn, profession, None::<i64>, None::<i64>])?;
-            written += 1;
-        }
-        drop(insert);
-        tx.commit()?;
-        Ok(written)
-    }
-}
-
-/// Parse a conquest objective FQN (`ach.conquests.<category>.<sub>...<leaf>`)
-/// into (category, subcategory, cadence). Cadence is `Some("weekly")` if the
-/// leaf ends with `_weekly` or path contains `.weekly.`, `Some("daily")` if
-/// the path contains `.daily.`, otherwise `None` for repeatable objectives.
-fn parse_conquest_fqn(fqn: &str) -> (String, Option<String>, Option<String>) {
-    // Expected shape: ach.conquests.<category>[.<subcategory>][...].<leaf>
-    let parts: Vec<&str> = fqn.split('.').collect();
-    if parts.len() < 4 || parts[0] != "ach" || parts[1] != "conquests" {
-        return ("unknown".to_string(), None, None);
-    }
-    let category = parts[2].to_string();
-    let subcategory = if parts.len() >= 5 {
-        Some(parts[3].to_string())
-    } else {
-        None
-    };
-
-    // Cadence: leaf-suffix or path-segment match.
-    let leaf = parts.last().copied().unwrap_or("");
-    let path_segments = &parts[..];
-    let cadence = if leaf.ends_with("_weekly") || path_segments.contains(&"weekly") {
-        Some("weekly".to_string())
-    } else if leaf.ends_with("_daily") || path_segments.contains(&"daily") {
-        Some("daily".to_string())
-    } else {
-        None
-    };
-
-    (category, subcategory, cadence)
 }
 
 impl Database {
-    /// Populate `conquest_objectives` from `ach.conquests.*` achievements.
-    /// Parses FQN segments to derive category, subcategory, and cadence.
-    pub fn populate_conquest_objectives(&self) -> Result<u64> {
-        let rows: Vec<(String, Option<u32>)> = {
-            let conn = self.conn.lock().unwrap();
-            let mut stmt = conn.prepare(
-                "SELECT fqn, string_id FROM objects WHERE kind = 'Achievement' AND fqn LIKE 'ach.conquests.%' AND is_canonical = 1",
-            )?;
-            let result: Vec<(String, Option<u32>)> = stmt
-                .query_map([], |row| {
-                    Ok((row.get::<_, String>(0)?, row.get::<_, Option<u32>>(1)?))
-                })?
-                .filter_map(|r| r.ok())
-                .collect();
-            drop(stmt);
-            result
-        };
-
-        let mut conn = self.conn.lock().unwrap();
-        let tx = conn.transaction()?;
-        let mut stmt = tx.prepare_cached(
-            "INSERT OR REPLACE INTO conquest_objectives (fqn, category, subcategory, cadence, string_id) VALUES (?1, ?2, ?3, ?4, ?5)",
-        )?;
-
-        let mut count = 0u64;
-        for (fqn, string_id) in &rows {
-            let (category, subcategory, cadence) = parse_conquest_fqn(fqn);
-            stmt.execute(rusqlite::params![
-                fqn,
-                category,
-                subcategory,
-                cadence,
-                string_id
-            ])?;
-            count += 1;
-        }
-
-        drop(stmt);
-        tx.commit()?;
-        Ok(count)
-    }
-
-    /// Populate `conquest_events` from the `cnqConquestInfoPrototype` singleton.
-    /// Each event is a CF40 8C7DAFE5 record: marker + length-prefixed ASCII
-    /// event name + inner sub-properties + a CC 0BAC73FD entry whose tail is
-    /// a length-prefixed `_pla_<planet>` string.
-    ///
-    /// 90 records total. record_size is bimodal: single-planet invasions
-    /// cluster at 68-80 bytes; themed/special events (Yavin, Onderon, Iokath,
-    /// Rishi, etc.) cluster at 7700-8400 bytes with many sub-bonuses. A
-    /// ~7400-byte empty gap separates the clusters, so any threshold inside
-    /// the gap classifies cleanly. Observed split: 74 invasion / 16 themed.
-    ///
-    /// The threshold sits in that gap (not just above invasion sizes) on
-    /// purpose: the final record has no following marker, so its measured
-    /// size runs to end-of-payload and absorbs the singleton's trailing CFE0
-    /// reference array. A tight threshold (e.g. 200) would mis-tag that one
-    /// inflated invasion record as themed; a gap-centered threshold cannot.
-    ///
-    /// Returns rows inserted.
-    pub fn populate_conquest_events(&self) -> Result<u64> {
-        self.flush()?;
-        use base64::{engine::general_purpose::STANDARD as BASE64, Engine as _};
-
-        const EVENT_MARKER: [u8; 4] = [0x8C, 0x7D, 0xAF, 0xE5];
-        const PLANET_CC: [u8; 4] = [0x0B, 0xAC, 0x73, 0xFD];
-        // Centered in the ~7400-byte gap between the invasion cluster (<=80B)
-        // and the themed cluster (>=7700B). Robust to last-record inflation.
-        const THEMED_SIZE_THRESHOLD: usize = 1000;
-
-        let payload: Option<Vec<u8>> = {
-            let conn = self.conn.lock().unwrap();
-            let row: Option<String> = conn
-                .query_row(
-                    "SELECT payload_b64 FROM singletons WHERE fqn = 'cnqConquestInfoPrototype'",
-                    [],
-                    |r| r.get(0),
-                )
-                .ok();
-            row.and_then(|b64| BASE64.decode(b64).ok())
-        };
-        let Some(payload) = payload else {
-            return Ok(0);
-        };
-
-        // Find every 8C7DAFE5 marker position.
-        let mut positions: Vec<usize> = Vec::new();
-        let mut i = 0;
-        while i + 9 <= payload.len() {
-            if payload[i] == 0xCF
-                && payload[i + 1] == 0x40
-                && payload[i + 2] == 0x00
-                && payload[i + 3] == 0x00
-                && payload[i + 5..i + 9] == EVENT_MARKER
-            {
-                positions.push(i);
-                i += 9;
-            } else {
-                i += 1;
-            }
-        }
-
-        let mut conn = self.conn.lock().unwrap();
-        let tx = conn.transaction()?;
-        let mut inserted = 0u64;
-        {
-            let mut insert = tx.prepare_cached(
-                "INSERT OR REPLACE INTO conquest_events \
-                    (ordinal, event_name, planet_code, event_kind, record_size) \
-                 VALUES (?1, ?2, ?3, ?4, ?5)",
-            )?;
-            for (ord, &pos) in positions.iter().enumerate() {
-                let end = positions.get(ord + 1).copied().unwrap_or(payload.len());
-                let record = &payload[pos..end];
-
-                // Event name: the first printable ASCII run of >= 4 chars
-                // after the 9-byte CF40 marker. The name is preceded by a
-                // `06 <len>` string tag, but we take the first qualifying run
-                // rather than trust the length prefix -- robust against the
-                // tag bytes themselves being non-printable separators.
-                let name = extract_ascii_strings(&record[9..], 4).into_iter().next();
-
-                // Planet code: find CC 0BAC73FD then the trailing
-                // length-prefixed `_pla_<planet>` ASCII string.
-                let planet = find_planet_code_after_cc(record, &PLANET_CC);
-
-                if let Some(name) = name {
-                    let kind = if record.len() >= THEMED_SIZE_THRESHOLD {
-                        "themed"
-                    } else {
-                        "invasion"
-                    };
-                    insert
-                        .execute(params![ord as i64, name, planet, kind, record.len() as i64,])?;
-                    inserted += 1;
-                }
-            }
-        }
-        tx.commit()?;
-        Ok(inserted)
-    }
-
-    /// Populate `conquest_schedule` from the `cnqSchedulePrototype` singleton:
-    /// the weekly conquest rotation. Each CF40 record (after the header)
-    /// carries an event reference (`cf`/`c7` tag + 8-byte GUID) and a week
-    /// ordinal encoded as a big-endian u16 immediately after the constant
-    /// `cb 01 ee fe bd 02 c9` anchor.
-    ///
-    /// The event GUID is resolved to a conquest event by locating those bytes
-    /// inside `cnqConquestInfoPrototype` and mapping the offset to the
-    /// enclosing CF40 8C7DAFE5 event record -- the same enumeration as
-    /// `populate_conquest_events`, so `event_ordinal` aligns with
-    /// `conquest_events.ordinal`. Returns rows inserted.
-    pub fn populate_conquest_schedule(&self) -> Result<u64> {
-        self.flush()?;
-        const EVENT_MARKER: [u8; 4] = [0x8C, 0x7D, 0xAF, 0xE5];
-        // Constant preamble in each schedule record, followed by the BE u16 week.
-        const WEEK_ANCHOR: [u8; 7] = [0xCB, 0x01, 0xEE, 0xFE, 0xBD, 0x02, 0xC9];
-
-        let Some(info) = self.load_singleton_payload("cnqConquestInfoPrototype") else {
-            return Ok(0);
-        };
-        let Some(sched) = self.load_singleton_payload("cnqSchedulePrototype") else {
-            return Ok(0);
-        };
-
-        // Event record start offsets, in order (ordinal == index, matching
-        // conquest_events). Each runs until the next start or end-of-payload.
-        let mut ev_starts: Vec<usize> = Vec::new();
-        {
-            let mut i = 0;
-            while i + 9 <= info.len() {
-                if info[i] == 0xCF
-                    && info[i + 1] == 0x40
-                    && info[i + 2] == 0
-                    && info[i + 3] == 0
-                    && info[i + 5..i + 9] == EVENT_MARKER
-                {
-                    ev_starts.push(i);
-                    i += 9;
-                } else {
-                    i += 1;
-                }
-            }
-        }
-        let event_name = |idx: usize| -> Option<String> {
-            let start = ev_starts[idx];
-            let end = ev_starts.get(idx + 1).copied().unwrap_or(info.len());
-            extract_ascii_strings(&info[start + 9..end], 4)
-                .into_iter()
-                .next()
-        };
-        // Map an 8-byte event GUID to the enclosing event record's ordinal.
-        let event_for_guid = |guid: &[u8]| -> Option<usize> {
-            let pos = info.windows(8).position(|w| w == guid)?;
-            match ev_starts.binary_search(&pos) {
-                Ok(i) => Some(i),
-                Err(0) => None,
-                Err(i) => Some(i - 1),
-            }
-        };
-
-        let mut conn = self.conn.lock().unwrap();
-        let tx = conn.transaction()?;
-        let mut inserted = 0u64;
-        {
-            let mut insert = tx.prepare_cached(
-                "INSERT OR REPLACE INTO conquest_schedule \
-                    (week_ordinal, event_guid, event_ordinal, event_name) \
-                 VALUES (?1, ?2, ?3, ?4)",
-            )?;
-            let positions = cf40_marker_positions(&sched);
-            for (idx, &pos) in positions.iter().enumerate() {
-                let end = positions.get(idx + 1).copied().unwrap_or(sched.len());
-                let record = &sched[pos..end];
-
-                // Event ref: first `cf` (non-marker) or `c7` after the 8-byte
-                // marker+hash, followed by an 8-byte GUID.
-                let mut guid: Option<&[u8]> = None;
-                let mut j = 8;
-                while j + 9 <= record.len() {
-                    let is_ref = (record[j] == 0xCF
-                        && !(record[j + 1] == 0x40 && record[j + 2] == 0 && record[j + 3] == 0))
-                        || record[j] == 0xC7;
-                    if is_ref {
-                        guid = Some(&record[j + 1..j + 9]);
-                        break;
-                    }
-                    j += 1;
-                }
-                let Some(guid) = guid else { continue };
-
-                // Week: BE u16 immediately after the constant anchor.
-                let Some(a) = record
-                    .windows(WEEK_ANCHOR.len())
-                    .position(|w| w == WEEK_ANCHOR)
-                else {
-                    continue;
-                };
-                let wk_pos = a + WEEK_ANCHOR.len();
-                if wk_pos + 2 > record.len() {
-                    continue;
-                }
-                let week = ((record[wk_pos] as i64) << 8) | record[wk_pos + 1] as i64;
-
-                let guid_hex: String = guid.iter().map(|b| format!("{b:02x}")).collect();
-                let ordinal = event_for_guid(guid);
-                let name = ordinal.and_then(event_name);
-                insert.execute(params![week, guid_hex, ordinal.map(|o| o as i64), name])?;
-                inserted += 1;
-            }
-        }
-        tx.commit()?;
-        Ok(inserted)
-    }
-
     /// Load and base64-decode a singleton payload by FQN. Returns None when
     /// the singleton is absent or its base64 fails to decode. Shared by the
     /// per-singleton prototype decoders.
@@ -1037,110 +563,6 @@ impl Database {
             .ok();
         drop(conn);
         b64.and_then(|b| BASE64.decode(b).ok())
-    }
-
-    /// Populate `armor_classes` from the `cbtArmorTablePrototype` singleton.
-    /// Each CF40 record carries a `02 <code>` byte and a `03 06 <len> <name>`
-    /// length-prefixed class name. Returns rows inserted.
-    pub fn populate_armor_classes(&self) -> Result<u64> {
-        self.flush()?;
-        let Some(payload) = self.load_singleton_payload("cbtArmorTablePrototype") else {
-            return Ok(0);
-        };
-
-        // CF40 record positions.
-        let positions = cf40_marker_positions(&payload);
-
-        let mut conn = self.conn.lock().unwrap();
-        let tx = conn.transaction()?;
-        let mut inserted = 0u64;
-        {
-            let mut insert = tx.prepare_cached(
-                "INSERT OR REPLACE INTO armor_classes (ordinal, code, name) VALUES (?1, ?2, ?3)",
-            )?;
-            for (ord, &pos) in positions.iter().enumerate() {
-                let end = positions.get(ord + 1).copied().unwrap_or(payload.len());
-                let record = &payload[pos..end];
-
-                // Name: the length-prefixed string after the `03 06` tag.
-                let name = find_length_prefixed_string(record, &[0x03, 0x06]);
-                // Code: the byte after the leading `02` tag (record[10]).
-                let code = record.get(10).copied();
-
-                if let (Some(name), Some(code)) = (name, code) {
-                    insert.execute(params![ord as i64, code as i64, name])?;
-                    inserted += 1;
-                }
-            }
-        }
-        tx.commit()?;
-        Ok(inserted)
-    }
-
-    /// Populate `stat_curve_values` from the `cbtShieldPerLevel` singleton.
-    /// It is a multi-record, float-dense table: each CF40 record carries one
-    /// or more typed floats (`04` tag + f32 LE). Every float is recorded in
-    /// payload order, grouped by the enclosing CF40 field-hash.
-    ///
-    /// These are LITERAL stored values with no level/stat semantics: the
-    /// curve is 2D (an undecoded segment key x level), so this is an honest
-    /// raw dump for downstream prose/chart rendering, separated by curve_hash.
-    ///
-    /// The other per-level singletons (cbtArmorPerLevel, cbtStandardRatingInfo,
-    /// chrGearScorePrototype) are single-record nested structures whose typed
-    /// stream is not cleanly float-only; they need dedicated grammar work and
-    /// are intentionally deferred so this table stays noise-free.
-    ///
-    /// Returns rows inserted.
-    pub fn populate_stat_curve_values(&self) -> Result<u64> {
-        self.flush()?;
-
-        const PROTOTYPE: &str = "cbtShieldPerLevel";
-        let Some(payload) = self.load_singleton_payload(PROTOTYPE) else {
-            return Ok(0);
-        };
-
-        let positions = cf40_marker_positions(&payload);
-
-        let mut conn = self.conn.lock().unwrap();
-        let tx = conn.transaction()?;
-        let mut inserted = 0u64;
-        {
-            // Idempotent on re-run: clear this prototype's rows first. The
-            // table has a surrogate PK with no natural-key collision, so a
-            // plain re-insert would otherwise accumulate duplicates when
-            // kessel runs against an existing output DB.
-            tx.execute(
-                "DELETE FROM stat_curve_values WHERE prototype = ?1",
-                params![PROTOTYPE],
-            )?;
-            let mut insert = tx.prepare_cached(
-                "INSERT INTO stat_curve_values (prototype, curve_hash, ordinal, value) \
-                 VALUES (?1, ?2, ?3, ?4)",
-            )?;
-            // Running ordinal per curve_hash.
-            let mut ordinals: std::collections::HashMap<String, i64> =
-                std::collections::HashMap::new();
-            for (idx, &pos) in positions.iter().enumerate() {
-                let end = positions.get(idx + 1).copied().unwrap_or(payload.len());
-                let record = &payload[pos..end];
-                // Field-hash: the 5 bytes after the `cf 40 00 00` marker.
-                let curve_hash = record
-                    .get(4..9)
-                    .map(|h| h.iter().map(|b| format!("{b:02x}")).collect::<String>());
-                // Every `04 <f32>` typed float in the record body (after the
-                // 9-byte marker+hash).
-                for value in typed_floats_in(&record[9.min(record.len())..]) {
-                    let key = curve_hash.clone().unwrap_or_default();
-                    let ord = ordinals.entry(key).or_insert(0);
-                    insert.execute(params![PROTOTYPE, curve_hash, *ord, value as f64])?;
-                    *ord += 1;
-                    inserted += 1;
-                }
-            }
-        }
-        tx.commit()?;
-        Ok(inserted)
     }
 
     pub fn stats(&self) -> Result<Stats> {
@@ -1472,52 +894,6 @@ mod tests {
     }
 
     #[test]
-    fn conquest_fqn_class_with_subcategory() {
-        let (cat, sub, cad) =
-            parse_conquest_fqn("ach.conquests.class.bounty_hunter.abilities.carbonize");
-        assert_eq!(cat, "class");
-        assert_eq!(sub.as_deref(), Some("bounty_hunter"));
-        assert_eq!(cad, None);
-    }
-
-    #[test]
-    fn conquest_fqn_location_with_planet() {
-        let (cat, sub, _) =
-            parse_conquest_fqn("ach.conquests.location.tatooine.complete_any_mission");
-        assert_eq!(cat, "location");
-        assert_eq!(sub.as_deref(), Some("tatooine"));
-    }
-
-    #[test]
-    fn conquest_fqn_weekly_suffix() {
-        let (_, _, cad) = parse_conquest_fqn("ach.conquests.crafting.craft_any_weekly");
-        assert_eq!(cad.as_deref(), Some("weekly"));
-    }
-
-    #[test]
-    fn conquest_fqn_weekly_segment_in_path() {
-        let (cat, _, cad) = parse_conquest_fqn(
-            "ach.conquests.galactic_seasons.priority_objectives.weekly.fp_vet_hutt",
-        );
-        assert_eq!(cat, "galactic_seasons");
-        assert_eq!(cad.as_deref(), Some("weekly"));
-    }
-
-    #[test]
-    fn conquest_fqn_daily_segment_in_path() {
-        let (_, _, cad) = parse_conquest_fqn(
-            "ach.conquests.galactic_seasons.priority_objectives.daily.heroics_out_rim",
-        );
-        assert_eq!(cad.as_deref(), Some("daily"));
-    }
-
-    #[test]
-    fn conquest_fqn_rejects_non_conquest() {
-        let (cat, _, _) = parse_conquest_fqn("ach.alliance.alliance_growth.specialists.x");
-        assert_eq!(cat, "unknown");
-    }
-
-    #[test]
     fn init_schema_is_idempotent_and_creates_all_tables() {
         let path = temp_db_path("schema_idem");
         let db = Database::with_grammar(&path, None).unwrap();
@@ -1552,172 +928,6 @@ mod tests {
         for t in ["npc_details", "schematic_details", "talent_details"] {
             assert!(names.contains(t), "missing typed-details table: {t}");
         }
-    }
-
-    #[test]
-    fn populate_conquest_events_classifies_and_resists_last_record_inflation() {
-        let path = temp_db_path("conquest_events");
-        let db = Database::with_grammar(&path, None).unwrap();
-        db.init_schema().unwrap();
-
-        // One synthetic event record: 9-byte CF40 8C7DAFE5 marker, then a
-        // `06 <len> <name>` string, then a CC 0BAC73FD planet block ending in
-        // a length-prefixed `_pla_<planet>` run with a non-printable
-        // terminator. Mirrors the real cnqConquestInfoPrototype layout.
-        fn event_record(name: &str, planet: &str) -> Vec<u8> {
-            let mut r = vec![0xCF, 0x40, 0x00, 0x00, 0x00, 0x8C, 0x7D, 0xAF, 0xE5];
-            r.push(0x06);
-            r.push(name.len() as u8);
-            r.extend_from_slice(name.as_bytes());
-            r.extend_from_slice(&[0xCC, 0x0B, 0xAC, 0x73, 0xFD]);
-            let pla = format!("_pla_{planet}");
-            r.push(0x06);
-            r.push(pla.len() as u8);
-            r.extend_from_slice(pla.as_bytes());
-            r.push(0x00);
-            r
-        }
-
-        let mut payload = Vec::new();
-        // Record A: small single-planet invasion.
-        payload.extend_from_slice(&event_record("Alpha", "alpha"));
-        // Record B: themed -- padded past the 1000-byte gap threshold.
-        let b_start = payload.len();
-        payload.extend_from_slice(&event_record("Beta", "beta"));
-        payload.resize(b_start + 1200, 0x00);
-        // Record C: small invasion, but it is the LAST record and is trailed
-        // by ~280 bytes mimicking the singleton's top-level CFE0 array. Its
-        // measured size (~320) exceeds the old 200 threshold yet stays well
-        // inside the bimodal gap -- so it must still classify as invasion.
-        payload.extend_from_slice(&event_record("Gamma", "gamma"));
-        payload.resize(payload.len() + 280, 0x00);
-
-        {
-            use base64::{engine::general_purpose::STANDARD as BASE64, Engine as _};
-            let conn = db.conn.lock().unwrap();
-            conn.execute(
-                "INSERT INTO singletons \
-                    (fqn, payload_size, payload_b64, string_count, cf_e0_count, cf_40_count, header_hex) \
-                 VALUES ('cnqConquestInfoPrototype', ?1, ?2, 0, 0, 3, '')",
-                params![payload.len() as i64, BASE64.encode(&payload)],
-            )
-            .unwrap();
-        }
-
-        let inserted = db.populate_conquest_events().unwrap();
-        assert_eq!(inserted, 3);
-
-        let conn = db.conn.lock().unwrap();
-        let row = |ord: i64| -> (String, String, Option<String>, i64) {
-            conn.query_row(
-                "SELECT event_name, event_kind, planet_code, record_size \
-                 FROM conquest_events WHERE ordinal = ?1",
-                params![ord],
-                |r| Ok((r.get(0)?, r.get(1)?, r.get(2)?, r.get(3)?)),
-            )
-            .unwrap()
-        };
-
-        let a = row(0);
-        assert_eq!(a.0, "Alpha");
-        assert_eq!(a.1, "invasion");
-        assert_eq!(a.2.as_deref(), Some("_pla_alpha"));
-
-        let b = row(1);
-        assert_eq!(b.0, "Beta");
-        assert_eq!(b.1, "themed");
-
-        // The crux: an inflated final invasion record must not become themed.
-        let c = row(2);
-        assert_eq!(c.0, "Gamma");
-        assert_eq!(c.1, "invasion");
-        assert_eq!(c.2.as_deref(), Some("_pla_gamma"));
-        assert!(
-            c.3 > 200,
-            "last record size {} must exceed the old 200 threshold to guard the fix",
-            c.3
-        );
-        assert!(
-            c.3 < 1000,
-            "last record size {} must stay within the gap",
-            c.3
-        );
-    }
-
-    #[test]
-    fn populate_conquest_schedule_resolves_weeks_to_events() {
-        let path = temp_db_path("conquest_schedule");
-        let db = Database::with_grammar(&path, None).unwrap();
-        db.init_schema().unwrap();
-
-        let guid_a: [u8; 8] = [0x07, 0x1d, 0x4c, 0xff, 0x6c, 0x65, 0xff, 0xfe];
-        let guid_b: [u8; 8] = [0x7c, 0x32, 0xfd, 0x18, 0x78, 0x30, 0x33, 0xab];
-        let guid_x: [u8; 8] = [0xde, 0xad, 0xbe, 0xef, 0x00, 0x11, 0x22, 0x33];
-
-        // Each conquest event record: CF40 8C7DAFE5 marker, 06 <len> <name>,
-        // then its identifying GUID embedded in the body.
-        fn event_record(name: &str, guid: &[u8; 8]) -> Vec<u8> {
-            let mut r = vec![0xCF, 0x40, 0x00, 0x00, 0x43, 0x8C, 0x7D, 0xAF, 0xE5];
-            r.push(0x06);
-            r.push(name.len() as u8);
-            r.extend_from_slice(name.as_bytes());
-            r.extend_from_slice(&[0x01, 0x01]);
-            r.extend_from_slice(guid);
-            r
-        }
-        let mut info = Vec::new();
-        info.extend_from_slice(&event_record("Alpha", &guid_a));
-        info.extend_from_slice(&event_record("Beta", &guid_b));
-        seed_singleton(&db, "cnqConquestInfoPrototype", &info);
-
-        // Each schedule record: marker + hash, 03 02, cf <guid>, then the
-        // week anchor `cb 01 ee fe bd 02 c9` + BE u16 week, then trailer.
-        fn sched_record(guid: &[u8; 8], week: u16) -> Vec<u8> {
-            let mut r = vec![0xCF, 0x40, 0x00, 0x00, 0x43, 0x8D, 0xCC, 0x77];
-            r.extend_from_slice(&[0x03, 0x02, 0xCF]);
-            r.extend_from_slice(guid);
-            r.extend_from_slice(&[0xCB, 0x01, 0xEE, 0xFE, 0xBD, 0x02, 0xC9]);
-            r.extend_from_slice(&week.to_be_bytes());
-            r.extend_from_slice(&[0x02, 0x02]);
-            r
-        }
-        let mut sched = Vec::new();
-        sched.extend_from_slice(&sched_record(&guid_a, 1001)); // -> Alpha
-        sched.extend_from_slice(&sched_record(&guid_b, 1002)); // -> Beta
-        sched.extend_from_slice(&sched_record(&guid_x, 1003)); // unresolved GUID
-        seed_singleton(&db, "cnqSchedulePrototype", &sched);
-
-        let n = db.populate_conquest_schedule().unwrap();
-        assert_eq!(n, 3);
-
-        let conn = db.conn.lock().unwrap();
-        let row = |wk: i64| -> (String, Option<i64>, Option<String>) {
-            conn.query_row(
-                "SELECT event_guid, event_ordinal, event_name FROM conquest_schedule \
-                 WHERE week_ordinal = ?1",
-                params![wk],
-                |r| Ok((r.get(0)?, r.get(1)?, r.get(2)?)),
-            )
-            .unwrap()
-        };
-        assert_eq!(
-            row(1001),
-            (
-                "071d4cff6c65fffe".to_string(),
-                Some(0),
-                Some("Alpha".to_string())
-            )
-        );
-        assert_eq!(
-            row(1002),
-            (
-                "7c32fd18783033ab".to_string(),
-                Some(1),
-                Some("Beta".to_string())
-            )
-        );
-        // Unresolved GUID: row exists, event_ordinal/name NULL.
-        assert_eq!(row(1003), ("deadbeef00112233".to_string(), None, None));
     }
 
     #[test]
@@ -1766,119 +976,6 @@ mod tests {
         b.extend_from_slice(&2.0f32.to_le_bytes());
         b.push(0x0a);
         assert_eq!(typed_floats_in(&b), vec![1.0, 2.0]);
-    }
-
-    #[test]
-    fn populate_armor_classes_decodes_code_and_name() {
-        let path = temp_db_path("armor_classes");
-        let db = Database::with_grammar(&path, None).unwrap();
-        db.init_schema().unwrap();
-
-        // Two records: marker(9) + 02 <code> 01 01 + 03 06 <len> <name>.
-        fn armor_record(hash5: &[u8; 5], code: u8, name: &str) -> Vec<u8> {
-            let mut r = vec![0xCF, 0x40, 0x00, 0x00];
-            r.extend_from_slice(hash5);
-            r.extend_from_slice(&[0x02, code, 0x01, 0x01]);
-            r.extend_from_slice(&[0x03, 0x06, name.len() as u8]);
-            r.extend_from_slice(name.as_bytes());
-            r
-        }
-        let mut payload = Vec::new();
-        payload.extend_from_slice(&armor_record(
-            &[0x48, 0xa2, 0xf9, 0x99, 0xa3],
-            0x5b,
-            "medium",
-        ));
-        payload.extend_from_slice(&armor_record(
-            &[0x48, 0xa2, 0xf9, 0x99, 0xa3],
-            0x5a,
-            "light",
-        ));
-
-        seed_singleton(&db, "cbtArmorTablePrototype", &payload);
-        let n = db.populate_armor_classes().unwrap();
-        assert_eq!(n, 2);
-
-        let conn = db.conn.lock().unwrap();
-        let (code, name): (i64, String) = conn
-            .query_row(
-                "SELECT code, name FROM armor_classes WHERE ordinal = 0",
-                [],
-                |r| Ok((r.get(0)?, r.get(1)?)),
-            )
-            .unwrap();
-        assert_eq!(code, 0x5b);
-        assert_eq!(name, "medium");
-        let light: String = conn
-            .query_row(
-                "SELECT name FROM armor_classes WHERE ordinal = 1",
-                [],
-                |r| r.get(0),
-            )
-            .unwrap();
-        assert_eq!(light, "light");
-    }
-
-    #[test]
-    fn populate_stat_curve_values_groups_floats_by_curve_hash() {
-        let path = temp_db_path("stat_curves");
-        let db = Database::with_grammar(&path, None).unwrap();
-        db.init_schema().unwrap();
-
-        // Two 17-byte shield records sharing one hash, then one with another.
-        fn shield_record(hash5: &[u8; 5], value: f32, idx: u8) -> Vec<u8> {
-            let mut r = vec![0xCF, 0x40, 0x00, 0x00];
-            r.extend_from_slice(hash5);
-            r.push(0x04);
-            r.extend_from_slice(&value.to_le_bytes());
-            r.extend_from_slice(&[idx, 0x0a, 0x01]);
-            r
-        }
-        let h1 = [0x27, 0x9b, 0x23, 0xb1, 0x2e];
-        let h2 = [0x27, 0x9b, 0x23, 0xb1, 0x2f];
-        let mut payload = Vec::new();
-        payload.extend_from_slice(&shield_record(&h1, 42.0, 0x02));
-        payload.extend_from_slice(&shield_record(&h1, 45.0, 0x03));
-        payload.extend_from_slice(&shield_record(&h2, 100.0, 0x02));
-
-        seed_singleton(&db, "cbtShieldPerLevel", &payload);
-        let n = db.populate_stat_curve_values().unwrap();
-        assert_eq!(n, 3);
-
-        // Idempotent on re-run: a second populate must not duplicate rows.
-        db.populate_stat_curve_values().unwrap();
-        {
-            let conn = db.conn.lock().unwrap();
-            let total: i64 = conn
-                .query_row("SELECT COUNT(*) FROM stat_curve_values", [], |r| r.get(0))
-                .unwrap();
-            assert_eq!(total, 3, "re-run must not duplicate stat_curve_values rows");
-        }
-
-        let conn = db.conn.lock().unwrap();
-        // First hash: ordinals 0,1 with values 42,45.
-        let vals: Vec<(i64, f64)> = {
-            let mut stmt = conn
-                .prepare(
-                    "SELECT ordinal, value FROM stat_curve_values \
-                     WHERE curve_hash = '279b23b12e' ORDER BY ordinal",
-                )
-                .unwrap();
-            stmt.query_map([], |r| Ok((r.get(0)?, r.get(1)?)))
-                .unwrap()
-                .map(Result::unwrap)
-                .collect()
-        };
-        assert_eq!(vals, vec![(0, 42.0), (1, 45.0)]);
-        // Second hash restarts ordinal at 0.
-        let (ord, val): (i64, f64) = conn
-            .query_row(
-                "SELECT ordinal, value FROM stat_curve_values WHERE curve_hash = '279b23b12f'",
-                [],
-                |r| Ok((r.get(0)?, r.get(1)?)),
-            )
-            .unwrap();
-        assert_eq!((ord, val), (0, 100.0));
     }
 
     #[test]
